+++ src/config.ts
@@ -1,6 +1,8 @@
-// export const baseUrlDevelopment = 'http://10.0.2.2:5015';
-export const baseUrlDevelopment = 'http://localhost:5015';
-export const baseUrlProduction = 'https://littleterrarium.one';
+// export const backendUrlDevelopment = 'http://10.0.2.2:5015';
+export const backendUrlDevelopment = 'http://localhost:5015';
+export const backendUrlProduction = 'https://littleterrarium.one';
+export const frontendUrlDevelopment = 'http://localhost:4200';
+export const frontendUrlProduction = 'https://littleterrarium.one';
 export const endpoint = '/api';
 export const theme = 'light';
 export const availableThemes = ['light', 'dark'];